+++ packages/atom-easy-jsdoc/CHANGELOG.md
@@ -1,3 +1,11 @@
+## 4.7.1
+- Fix issue with null default values
+
+## 4.7.0
+- Add class support
+- Add async support
+- Improve error messages
+
 ## 4.6.0
 - Split JSDoc rendering and function parsing to improve test coverage.
 - Change from acorn to babylon parser to get better real code support.